+++ app/router.ts
@@ -95,7 +95,6 @@ const Router = EmberRouter.extend({
 Router.map(function() {
     // All non-guid routes (except error routes) belong above "Guid Routing"
     this.route('home', { path: '/' });
-    this.route('old-home');
     this.route('dashboard');
     this.route('goodbye');
     this.route('institutions');