+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Changed
 - `new-home` route is now `home`
-- `home` route is now `old-home`
+
+### Removed
+- `old-home` route
+- `old-home` tests
 
 ## [19.5.1] - 2019-06-24
 ### Added