+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- `new-home` route is now `home`
+
+### Removed
+- `old-home` route
+- `old-home` tests
 
 ## [19.5.1] - 2019-06-24
 ### Added